requirements.txt
discord.py
pysocks
requests
mcstatus